Create static_code_analyzer.py
A project were you in stages implement more and more checks from the PEP8 guidelines. A bit of a mismatch of processes at the moment, as each stage had different requirements. Might work on it further with the ast module to replace some of the stuff from the earlier stages, to make it more uniform.