Create bus_company.py
A problem from JetBrains academy where the task was split into different parts where each part posed a different problem to solve, with no regards to former parts in the 
task. So every function is meant to solve a new problem and therefore I decided to experiment with different approaches.